loadbalancer/requirements.txt
requests==2.32.3
PyYAML==6.0.2
Jinja2==3.1.6
